refactor: resource server token introspection request process
